modifications

+++ controllers/notes.js
@@ -11,19 +11,46 @@ const getTokenFrom = request => {
   return null
 }
 
-notesRouter.get('/', async (request, response) => {
-  const notes = await Note
-    .find({}).populate('user', { username: 1, name: 1 })
+notesRouter.get('/get-users-notes/:id', async (request, response) => {
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
 
-  response.json(notes.map(note => note.toJSON()))
-})
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
 
-notesRouter.put('/:id', (request, response, next) => {
+  if (
+    decodedToken.id.toString() === request.params.id.toString() ||
+    decodedToken.admin === true
+  ) {
+    const notes = await Note.find({ user: request.params.id })
+      .populate('user', { username: 1, name: 1 })
+      .populate('noteStack', { title: 1 });
+    response.json(notes.reverse().map(note => note.toJSON));
+  }
+});
+
+notesRouter.put('/:id', async (request, response, next) => {
   const body = request.body
+  const token = getTokenFrom(request);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  
+  const note = await NoteStack.findById(request.params.id)
+  if (note) {
+    response.json(note.toJSON())
+  } else {
+    response.status(404).end()
+  }
 
   const note = {
+    title: body.title,
     content: body.content,
-    important: body.important,
+    noteStack: noteStack,
+    lastModifiedDate: new Date(),
   }
 
   Note.findByIdAndUpdate(request.params.id, note, { new: true })
@@ -34,6 +61,13 @@ notesRouter.put('/:id', (request, response, next) => {
 })
 
 notesRouter.get('/:id', async (request, response) => {
+  const token = getTokenFrom(request);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+
   const note = await Note.findById(request.params.id)
   if (note) {
     response.json(note.toJSON())
@@ -41,7 +75,6 @@ notesRouter.get('/:id', async (request, response) => {
     response.status(404).end()
   }
 })
-
 
 notesRouter.post('/', async (request, response) => {
   const body = request.body
@@ -56,9 +89,11 @@ notesRouter.post('/', async (request, response) => {
 
   const note = new Note({
     content: body.content,
-    important: body.important === undefined ? false : body.important,
-    date: new Date(),
-    user: user._id
+    createdDate: new Date(),
+    lastModifiedDate: new Date(),
+    user: user._id,
+    noteStack: body.noteStack,
+    title: body.title,
   })
 
   const savedNote = await note.save()
@@ -69,6 +104,13 @@ notesRouter.post('/', async (request, response) => {
 })
 
 notesRouter.delete('/:id', async (request, response) => {
+  const token = getTokenFrom(request)
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   await Note.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
